Recommendation System
Recommender systems are beneficial to both the service providers and
users. They Reduce the transaction costs of finding and selecting items
in an online shopping environment. Recommendation systems have also
proved to improve decision making process and quality.
With the increase in number of people who are very much interested in
listening to the music, it is highly advantage for the sites to
introduce a recommendation technique which would recommend songs to the
user which more likely to the user interest. For this System Pearson and
Cosine Similarity techniques used along with User based collaborative
filtering.